+++ friendlysam/model/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from friendlysam.model.core import Part, InsanityError
-from friendlysam.model.nodes import Node, Cluster, Storage, ResourceNetwork
+from friendlysam.model.nodes import Node, Cluster, Storage, FlowNetwork